Moved expression parser tests to their own file
