+++ tests/test_expression_parser.h
@@ -1,0 +1,10 @@
+#ifndef ZN_TEST_EXPRESSION_PARSER_H
+#define ZN_TEST_EXPRESSION_PARSER_H
+
+namespace zylann::tests {
+
+void test_expression_parser();
+
+} // namespace zylann::tests
+
+#endif // ZN_TEST_EXPRESSION_PARSER_H
